Use a IoU sum (instead of average) in order to decide whether we should accept or refuse alternative chopsticks. Fix a bug that caused hidden or undetected tips to have a wrong position because they weren't compensated for camera motion. Make VideoFramePainters more flexible.

+++ src/model/tracking/Tip.hpp
@@ -26,8 +26,8 @@ namespace model {
                 boost::circular_buffer<TrackingStatus> recentTrackingStatuses,
                 int nbDetectionsAsBigTip,
                 int nbDetectionsAsSmallTip,
-                int x, int y,
-                int width, int height) :
+                double x, double y,
+                double width, double height) :
                     id(id),
                     recentShapes(recentShapes),
                     recentTrackingStatuses(recentTrackingStatuses),
